pyproject.toml
[project]
name = "google-drive-connector"
version = "0.1.0"
description = "Google Drive connector for Databricks - ingest files into Unity Catalog"
requires-python = ">=3.10"
dependencies = [
    "databricks-sdk>=0.40.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.9.0",
    "pytest>=8.0.0",
]

[tool.ruff]
line-length = 120
